Updated Week2 lab with hints, more examples
Based on running this lab in January 2018, students really struggled with, e.g., what a list() or matrix() look like and how they should be constructed because the output list and matrix were so large. When constructed from a vector from the original rent.US data frame, they spill over into the console and its harder for students to see how they work.